Refactored the code to make it object oriented and organized
