+++ skema/text_reading/model_linker/walkers/petrinet.py
@@ -1,0 +1,9 @@
+from typing import Optional, Any
+
+from walkers.json import JsonDictWalker
+
+
+class PetriNetWalker(JsonDictWalker):
+
+    def _filter(self, obj_name: Optional[str], obj: Any, index: Optional[int]) -> bool:
+        return obj_name in {"states", "transitions"}
